% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{cal_cor}
\alias{cal_cor}
\title{cal_cor}
\usage{
cal_cor(data, ceLength = 10, noise = 200, min_CE = NA, max_CE = NA)
}
\arguments{
\item{data}{A tibble.}

\item{ceLength}{ceLength.}

\item{noise}{noise.}
}
\value{
A tibble.
}
\description{
calculate correlation df.
}
\examples{
data(CA)
df <- cal_cor(CA, max_CE = 35)
}
