% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{paraSet}
\alias{paraSet}
\title{paraSet}
\usage{
paraSet(df, corThreshold = 0.8)
}
\arguments{
\item{df}{A tibble like demo}

\item{corThreshold}{Threshold of correlation}
}
\value{
A tibble.
}
\description{
Set parameters for df
}
\examples{
df <- paraSet(df = df)
}
