% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{plotMultiMS}
\alias{plotMultiMS}
\title{plotMultiMS}
\usage{
plotMultiMS(df, scale_size = c(4, 16), stroke = 1)
}
\arguments{
\item{df}{A tibble like demo.}
}
\value{
A ggplot2 object.
}
\description{
plotMultiMS.
}
\examples{
plotMultiMS(df = df, scale_size = c(2, 10))
}
