% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{plotMultiMS}
\alias{plotMultiMS}
\title{plotMultiMS}
\usage{
plotMultiMS(df)
}
\arguments{
\item{df}{A tibble like demo.}
}
\value{
A ggplot2 object.
}
\description{
plotMultiMS.
}
\examples{
plotMultiMS(df = demo)
}
